app/requirements.txt
mysql-connector-python
configparser
flask
